# BACKEND/API/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "GPAC REST API",
    "description": "Desarrollo de una API REST para una aplicación de gestión de tareas que ayude a los usuarios a organizar y administrar sus actividades diarias. La aplicación debe permitir a los usuarios crear, seguir y completar tareas, establecer fechas límite, asignar prioridades y añadir notas según sea necesario. Además, los usuarios pueden categorizar las tareas en proyectos o etiquetas para una mejor organización. La aplicación debe ofrecer funcionalidades de búsqueda y filtrado, permitiendo a los usuarios buscar tareas específicas y aplicar filtros según criterios como fecha, prioridad o proyecto."
  },
  "host": "localhost:3000",
  "basePath": "/backend",
  "schemes": [
    "http"
  ],
  "paths": {
    "/users/Login": {
      "post": {
        "description": "Method used to login into the app.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/AddUser": {
      "post": {
        "description": "Method used to creade a new user. All parameters are required.",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "adminPrivileges": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/users/SearchUsers": {
      "post": {
        "description": "Method used to get a list of users. It can be filtered and ordered (these parameters are optional).",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "filter": {
                  "example": "any"
                },
                "order": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/users/GetUsers": {
      "get": {
        "description": "Method used to get a list of users.",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/users/UpdateUser": {
      "put": {
        "description": "Method used to update a user's information. Parameter '_id' is required, parameters 'adminPrivileges' and 'password' are optional.",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "adminPrivileges": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "_id": {
                  "example": "XXXXXXXXXXXXXXXXXXXXXXXX"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/users/DeleteUser": {
      "delete": {
        "description": "Method used to delete a user from the app. Parameter '_id' is required.",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "example": "XXXXXXXXXXXXXXXXXXXXXXXX"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  }
}